Issues with decimal numbers in duration

+++ src/ReadingRoom.Umbraco.CloudflareStream/Models/CloudflareStreamMediaValue.cs
@@ -10,7 +10,7 @@ public class CloudflareStreamMediaValue : PublishedContentWrapped
         long size,
         long width,
         long height,
-        long duration,
+        double duration,
         DateTime uploadDate,
         CloudflareStreamPlaybackUrls playback,
         IPublishedValueFallback publishedValueFallback) : base(content, publishedValueFallback)
@@ -29,7 +29,7 @@ public class CloudflareStreamMediaValue : PublishedContentWrapped
     public long Height { get; }
     public long Width { get; }
     public string VideoId { get; }
-    public long Duration { get; }
+    public double Duration { get; }
     public DateTime UploadDate { get; }
 
     public string? AspectRatio => GetAspectRatio();
